+++ CHANGELOG.md
@@ -8,6 +8,16 @@ and this project adheres to
 
 ## [Unreleased]
 
+### Changed
+
+- Step Sequencer: Edit current clip instead of creating a new one.
+- Step Sequencer: Changing number of note per measure does not loop accross options anymore.
+
+### Fixed
+
+- Sequencer had a length of 0 on loading. It's now filled by default: 4 measures of 4 notes.
+- When changing number of notes per measures, the midi clip stays the same instead of time stretched.
+
 ## [0.6.0] - 2025-11-20
 
 ### Added