+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to
 
 ## [Unreleased]
 
+## [0.7.0] - 2026-01-16
+
 ### Added
 
 - Plugins: Distortion effect.
@@ -88,7 +90,13 @@ and this project adheres to
 - Initial release
 
 [unreleased]:
-  https://github.com/FundamentalFrequency/LMN-3-DAW/compare/v0.4.0...HEAD
+  https://github.com/FundamentalFrequency/LMN-3-DAW/compare/v0.7.0...HEAD
+[0.7.0]:
+  https://github.com/FundamentalFrequency/LMN-3-DAW/compare/v0.6.0...v0.7.0
+[0.6.0]:
+  https://github.com/FundamentalFrequency/LMN-3-DAW/compare/v0.5.0...v0.6.0
+[0.5.0]:
+  https://github.com/FundamentalFrequency/LMN-3-DAW/compare/v0.4.0...v0.5.0
 [0.4.0]:
   https://github.com/FundamentalFrequency/LMN-3-DAW/compare/v0.3.0...v0.4.0
 [0.3.0]: